[NOW] Code style cosmetic cleanups
